pyproject.toml
[tool.poetry]
name = "DHDRNet"
version = "0.1.0"
description = ""
authors = ["Shane Segal <shane@smsegal.ca>"]

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.4.0"
torchvision = "^0.5.0"
networkx = "^2.4"
colour-science = {extras = ["plotting", "optional"], version = "^0.3.15"}
colour-hdri = {extras = ["optional"], version = "^0.1.6"}
opencv-python = "^4.1.2"
matplotlib = "^3.1.3"
imageio = "^2.6.1"
black = "^19.10b0"
scikit-image = "^0.16.2"
more-itertools = "^8.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.11.1"
pylama = "^7.7.1"
ipykernel = "^5.1.4"
jupyter = "^1.0.0"
isort = "^4.3.21"
nbstripout = "^0.3.7"
pyflakes = "^2.1.1"
deprecated = "^1.2.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
